recommender-ai-service/requirements.txt
django
djangorestframework
requests
scikit-learn
numpy
